Run the AMQP to ZeroMQ bridge with "fedora-messaging consume"
Since the AMQP-> ZeroMQ bridge is just a consumer, run it with the
fedora-messaging CLI. To run the bridge, just do:

    fedora-messaging consume --callback="fedmsg_migration_tools.bridges:amqp_to_zmq"

On the one hand it's a bit confusing to use two different CLIs for the
bridges, but on the other hand we already have a CLI for this. Since
we're going to be providing systemd units to run it, though, I think
this will be okay.

Signed-off-by: Jeremy Cline <jcline@redhat.com>