Set the fedora-message body to the fedmsg "msg" key
AMQP publishers don't wrap everything in a "msg" key and the intention
was for the ZMQ->AMQP bridge to exactly mimic those messages. However,
the entire fedmsg was used as the body. This sets the body to the value
of the fedmsg's "msg" key and uses the value of the fedmsg's "headers"
key as the headers.

Fixes #20

Signed-off-by: Jeremy Cline <jcline@redhat.com>